クラスごとの責務が適切となるように、メソッドの定義を移行
