+++ 06.bowling_object/bowling_object.rb
@@ -4,5 +4,7 @@
 
 require './game'
 
+require 'debug'
+
 game = Game.new(ARGV[0])
 puts game.total_score
